+++ intelmq/tests/bots/experts/cymru_whois/test_expert.py
@@ -13,7 +13,7 @@ EXAMPLE_INPUT = {"__type": "Event",
 EXAMPLE_OUTPUT = {"__type": "Event",
                   "source.ip": "93.184.216.34",
                   "source.geolocation.cc": "EU",
-                  "source.registry": "ripencc",
+                  "source.registry": "RIPE",
                   "source.network": "93.184.216.0/24",
                   "source.allocated": "2008-06-02T00:00:00+00:00",
                   "source.asn": 15133,
@@ -26,7 +26,7 @@ EXAMPLE_INPUT6 = {"__type": "Event",
                   }
 EXAMPLE_OUTPUT6 = {"__type": "Event",
                    "destination.ip": "2001:500:88:200::8",  # iana.org
-                   "destination.registry": "arin",
+                   "destination.registry": "ARIN",
                    "destination.allocated": "2010-02-18T00:00:00+00:00",
                    "destination.as_name": "ICANN-DC - ICANN, US",
                    "destination.geolocation.cc": "US",
@@ -41,7 +41,7 @@ UNICODE_INPUT = {"__type": "Event",  # http://bgp.he.net/AS28333#_prefixes
 UNICODE_OUTPUT = {"__type": "Event",
                   "destination.ip": "189.50.192.1",
                   "time.observation": "2015-01-01T00:00:00+00:00",
-                  "destination.registry": "lacnic",
+                  "destination.registry": "LACNIC",
                   "destination.allocated": "2007-08-10T00:00:00+00:00",
                   "destination.as_name": "Local Datacenter Solu\xe7\xf5es em"
                                          " Comunica\xe7\xe3o Ltda., BR",
@@ -64,7 +64,7 @@ NO_ASN_OUTPUT = {"__type": "Event",
                  "source.geolocation.cc": 'RU',
                  "source.ip": '212.92.127.126',
                  "source.network": '212.92.127.0/24',
-                 "source.registry": 'ripencc',
+                 "source.registry": 'RIPE',
                  }