+++ intelmq/tests/lib/test_harmonization.py
@@ -251,6 +251,35 @@ class TestHarmonization(unittest.TestCase):
         self.assertTrue(harmonization.LowercaseString.is_valid(b'fooBar',
                                                                sanitize=True))
 
+    def test_registry_valid(self):
+        """ Test Registry.is_valid with valid arguments. """
+        self.assertTrue(harmonization.Registry.is_valid('RIPE'))
+        self.assertTrue(harmonization.Registry.is_valid('lacnic', sanitize=True))
+        self.assertTrue(harmonization.Registry.is_valid('RIPE-NCC', sanitize=True))
+        self.assertTrue(harmonization.Registry.is_valid('RIPENCC', sanitize=True))
+
+    def test_registry_invalid(self):
+        """ Test Registry.is_valid with invalid arguments. """
+        self.assertFalse(harmonization.Registry.is_valid('RIPE-NCC', sanitize=False))
+        self.assertFalse(harmonization.Registry.is_valid('RIPENCC', sanitize=False))
+
+    def test_registry_sanitize(self):
+        """ Test Registry.sanitize with valid arguments. """
+        self.assertEqual(harmonization.Registry.sanitize('ripe-ncc'), 'RIPE')
+
+    def test_uppercasestring_valid(self):
+        """ Test UppercaseString.is_valid with valid arguments. """
+        self.assertTrue(harmonization.UppercaseString.is_valid('FOOBAR'))
+
+    def test_uppercasestring_invalid(self):
+        """ Test UppercaseString.is_valid with invalid arguments. """
+        self.assertFalse(harmonization.UppercaseString.is_valid('fooBar'))
+
+    def test_uppercasestring_sanitize(self):
+        """ Test UppercaseString.sanitize with valid arguments. """
+        self.assertTrue(harmonization.UppercaseString.is_valid(b'fooBar',
+                                                               sanitize=True))
+
     def test_url_valid(self):
         """ Test URL.is_valid with valid arguments. """
         self.assertTrue(harmonization.URL.is_valid('http://example.com'))